refactor: seperate gstr_1 utilities
(cherry picked from commit c243e7ac92c271ef92f357017ee2632aa2978810)
